+++ backend/app/controllers/uoas.js
@@ -6,6 +6,7 @@ var client = require('app/db_bootstrap'),
     vl = require('validator'),
     UnitOfAnalysis = require('app/models/uoas'),
     UnitOfAnalysisType = require('app/models/uoatypes'),
+    UnitOfAnalysisTagLink = require('app/models/uoataglinks'),
     AccessMatrix = require('app/models/access_matrices'),
     Translation = require('app/models/translations'),
     Language = require('app/models/languages'),
@@ -114,6 +115,10 @@ module.exports = {
     deleteOne: function (req, res, next) {
         var thunkQuery = req.thunkQuery;
         co(function* () {
+            var result = yield thunkQuery(UnitOfAnalysisTagLink.select().where(UnitOfAnalysisTagLink.uoaId.equals(req.params.id)));
+            if (_.first(result)) {
+                throw new HttpError(403, 'Subject used in Subject to Tag link. Could not delete Subject');
+            }
             return yield thunkQuery(UnitOfAnalysis.delete().where(UnitOfAnalysis.id.equals(req.params.id)));
         }).then(function () {
             bologger.log({
